Analysis
Add: Feature to change the project days and update the graph

+++ MobileProjectEstimator/app/src/main/java/com/mobileprojectestimator/mobileprojectestimator/DataObjects/Project/Project.java
@@ -40,7 +40,7 @@ public class Project implements Serializable
     private String iconId;
     private boolean isDeleted;
     private boolean isTerminated;
-    private int finalPersonDays;
+    private double finalPersonDays;
     private double evaluatedPoints;
 
     /**
@@ -598,12 +598,12 @@ public class Project implements Serializable
         this.isTerminated = isTerminated;
     }
 
-    public int getFinalPersonDays()
+    public double getFinalPersonDays()
     {
         return finalPersonDays;
     }
 
-    public void setFinalPersonDays(int finalPersonDays)
+    public void setFinalPersonDays(double finalPersonDays)
     {
         this.finalPersonDays = finalPersonDays;
     }
